requirements.txt
aniso8601 == 8.0.0
click == 7.1.2
Flask == 1.1.2
Flask-RESTful == 0.3.8
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
flake8==3.8.4
pytest==6.1.2
pytest-cov==2.10.1
pytest-flask==1.0.0
pytest-xdist==2.1.0
gunicorn==20.0.4
psycopg2==2.8.6
pylint-flask==0.6
flask-bcrypt==0.7.1
azure-storage-blob==12.6.0